+++ trunk/Baselet/src/com/umlet/element/relation/DoubleStroke.java
@@ -8,7 +8,7 @@ public class DoubleStroke implements Stroke {
 	private Stroke outerStroke;
 	private Stroke innerStroke;
 
-	public DoubleStroke(int lineWidth, int distanceBetweenLines, int cap, int join, float miterlimit, float[] dash, float dash_phase) {
+	public DoubleStroke(float lineWidth, float distanceBetweenLines, int cap, int join, float miterlimit, float[] dash, float dash_phase) {
 		this.outerStroke = new BasicStroke(lineWidth + distanceBetweenLines, cap, join, miterlimit, dash, dash_phase);
 		this.innerStroke = new BasicStroke(lineWidth, cap, join, miterlimit, dash, dash_phase);
 	}
